refactor: Spinner 컴포넌트 관련 디렉토리 /client/components 하위에서 /client/components/@atom 하위로로 이동
